Split presto-session-property-managers module
Currently, the presto-session-property-managers module has gotten large due to the different implementations for file based
and db based session property managers. If implemented, this commit will split the module into 3,
``presto-session-property-managers-common`` will contain the API implementation for a session property manager
along with some tests. ``presto-file-session-property-manager`` will contain the file based implementation, and
``presto-db-session-property-manager`` will contain the db based implementation.
