+++ presto-db-session-property-manager/src/main/java/com/facebook/presto/session/db/DbSpecsProvider.java
