+++ presto-session-property-managers-common/src/main/java/com/facebook/presto/session/AbstractSessionPropertyManager.java
