+++ presto-session-property-managers-common/src/test/java/com/facebook/presto/session/AbstractTestSessionPropertyManager.java
