--- presto-session-property-managers/src/test/java/com/facebook/presto/session/file/TestFileSessionPropertyManagerConfig.java
@@ -1,47 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.session.file;
-
-import com.google.common.collect.ImmutableMap;
-import org.testng.annotations.Test;
-
-import java.io.File;
-import java.util.Map;
-
-import static com.facebook.airlift.configuration.testing.ConfigAssertions.assertFullMapping;
-import static com.facebook.airlift.configuration.testing.ConfigAssertions.assertRecordedDefaults;
-import static com.facebook.airlift.configuration.testing.ConfigAssertions.recordDefaults;
-
-public class TestFileSessionPropertyManagerConfig
-{
-    @Test
-    public void testDefaults()
-    {
-        assertRecordedDefaults(recordDefaults(FileSessionPropertyManagerConfig.class)
-                .setConfigFile(null));
-    }
-
-    @Test
-    public void testExplicitPropertyMappings()
-    {
-        Map<String, String> properties = new ImmutableMap.Builder<String, String>()
-                .put("session-property-manager.config-file", "/test.json")
-                .build();
-
-        FileSessionPropertyManagerConfig expected = new FileSessionPropertyManagerConfig()
-                .setConfigFile(new File("/test.json"));
-
-        assertFullMapping(properties, expected);
-    }
-}
