+++ presto-spark-package/pom.xml
@@ -152,7 +152,15 @@
 
         <dependency>
             <groupId>com.facebook.presto</groupId>
-            <artifactId>presto-session-property-managers</artifactId>
+            <artifactId>presto-file-session-property-manager</artifactId>
+            <version>${project.version}</version>
+            <type>zip</type>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.facebook.presto</groupId>
+            <artifactId>presto-db-session-property-manager</artifactId>
             <version>${project.version}</version>
             <type>zip</type>
             <scope>provided</scope>
